% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart_grouped.R
\name{bar_chart_grouped}
\alias{bar_chart_grouped}
\title{Generates grouped horizontal barchart with scenario triangles.}
\usage{
bar_chart_grouped(
  data,
  cat,
  foreground,
  background,
  triangles = NULL,
  series_labels,
  df_styles = NULL
)
}
\arguments{
\item{data}{data frame containing data to be plotted}

\item{cat}{vector cointaining category names of values}

\item{series_labels}{vector containing names of series to be shown on the plot}

\item{df_styles}{optional data frame containing styles of bars. It is necessary that df_style has the same column names as defined in series vector.}

\item{series}{vector containing names of columns in data with values to plot}
}
\value{
SVG string containing chart
}
\description{
Generates grouped horizontal barchart with scenario triangles.
}
\examples{

#preparing data frames
data <- data.frame(
city = c("Berlin", "Paris", "London", "Munich", "Vienna"),
AC = c(592, 1166, 618, 795, 538),
PL = c(570, 950, 800, 780, 460),
triangles = c(545, 800, 900, 600, 538) #AC toten bardziej na wierzchu
)

#preparing the styles data frame
df_styles <- data.frame(
 AC = c("actual","actual","actual","actual","actual"),
 PL = c("plan","plan","plan","plan","plan"),
 triangles = c("previous", "previous","previous","previous","previous"))

#creating the svg string
barchart_grouped <- bar_chart_grouped(data, data$city, c("triangles", "AC", "PL"), c("triangles", "AC", "PL"), df_styles)

#showing the plot
barchart_grouped \%>\% SVGrenderer()

}
