% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart.R
\name{bar_chart_relative_variance}
\alias{bar_chart_relative_variance}
\title{Generate bar chart with relative variance (in percents)}
\usage{
bar_chart_relative_variance(
  cat,
  baseline,
  real,
  colors = 1,
  data = NULL,
  y_title,
  y_style = "previous"
)
}
\arguments{
\item{baseline}{vector containing base values or name of column in data with base values}

\item{real}{vector containing values that will be compared to baseline  or name of column in data with that values}

\item{colors}{1 if green color represents positive values having good buisness impact and red negative values having bad impact or 2 if otherwise}

\item{data}{data frame with columns containing data for x, baseline or real series}

\item{y_title}{title of the series values}

\item{y_style}{style of y axis to indicate baseline scenario}
}
\value{
SVG string containing chart
}
\description{
Generate bar chart with relative variance (in percents)
}
