% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart_plot_grouped.R
\name{barchart_plot_grouped}
\alias{barchart_plot_grouped}
\title{Generates grouped horizontal barchart with scenario triangles.}
\usage{
barchart_plot_grouped(data, cat, series, series_labels, df_styles = NULL)
}
\arguments{
\item{data}{data frame containing data to be plotted}

\item{cat}{vector cointaining category names of values}

\item{series}{vector containing names of columns in data with values to plot}

\item{series_labels}{vector containing names of series to be shown on the plot}

\item{df_styles}{optional data frame containing styles of bars. It is necessary that df_style has the same column names as defined in series vector.}
}
\value{
SVG string containing chart
}
\description{
Generates grouped horizontal barchart with scenario triangles.
}
