% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart_plot.R
\name{barchart_plot_index}
\alias{barchart_plot_index}
\title{Generates basic horizontal barchart with index on a given value. If more than one series is supplied, stacked barchart is generated.}
\usage{
barchart_plot_index(data, cat, series, index_val, series_labels, styles = NULL)
}
\arguments{
\item{data}{data frame containing data to be plotted}

\item{cat}{vector cointaining category names of values}

\item{series}{vector containing names of columns in data with values to plot}

\item{index_val}{numeric value of the index}

\item{series_labels}{vector containing names of series to be shown on the plot}

\item{styles}{optional vector with styles of bars}
}
\value{
SVG string containing chart
}
\description{
Generates basic horizontal barchart with index on a given value. If more than one series is supplied, stacked barchart is generated.
}
