% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart_plot.R
\name{barchart_plot_relative_variance}
\alias{barchart_plot_relative_variance}
\title{Generate bar chart with relative variance (in percents)}
\usage{
barchart_plot_relative_variance(
  cat,
  baseline,
  real,
  colors = 1,
  y_title,
  y_style = "previous"
)
}
\arguments{
\item{cat}{vector containing category names of values}

\item{baseline}{vector containing base values}

\item{real}{vector containing values that will be compared to baseline}

\item{colors}{1 if green color represents positive values having good buisness impact and red negative values having bad impact or 2 if otherwise}

\item{y_title}{title of the series values}

\item{y_style}{style of y axis to indicate baseline scenario}
}
\value{
SVG string containing chart
}
\description{
Generate bar chart with relative variance (in percents)
}
