% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart_plot_waterfall.R
\name{barchart_plot_waterfall}
\alias{barchart_plot_waterfall}
\title{Generate horizontal waterfall chart}
\usage{
barchart_plot_waterfall(y, values, add_result = FALSE, result_title = NULL)
}
\arguments{
\item{y}{vector containing labels for y axis}

\item{values}{vector containing values that will be plotted}

\item{add_result}{boolean value if result bar should be plotted}

\item{result_title}{the title for the result bar. Ignored if add_result is false}
}
\value{

}
\description{
Generate horizontal waterfall chart
}
