% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_chart_functions.R
\name{column_chart}
\alias{column_chart}
\title{Generate basic column chart. If more than one series is supplied, stacked column plot is generated}
\usage{
column_chart(df, x, series = NULL, styles = NULL, interval = "months")
}
\arguments{
\item{df}{data frame in wide format containing data to be plotted}

\item{x}{vector containing labels for x axis or name of column in df with values of x axis labels}

\item{series}{vector containing names of columns in df with values to plot}

\item{styles}{optional vector with styles of bars}
}
\value{
SVG string containing chart
}
\description{
Generate basic column chart. If more than one series is supplied, stacked column plot is generated
}
