% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_chart_functions.R
\name{column_chart_absolute_variance}
\alias{column_chart_absolute_variance}
\title{Generate column chart with absolute variance}
\usage{
column_chart_absolute_variance(x, baseline, real, colors = 1, x_title = "PY")
}
\arguments{
\item{x}{vector containing labels for x axis}

\item{baseline}{vector containing base values}

\item{real}{vector containing values that will be compared to baseline}

\item{colors}{1 if green color represents positive values having good buisness impact and red negative values having bad impact or 2 if otherwise}

\item{x_title}{the title of the plot}
}
\value{
SVG string containing chart
}
\description{
Generate column chart with absolute variance
}
