% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_chart_functions.R
\name{column_chart_grouped}
\alias{column_chart_grouped}
\title{Generate grouped column chart for visualizing up to 3 data series}
\usage{
column_chart_grouped(
  x,
  foreground,
  background,
  triangles = NULL,
  titles,
  styles = NULL,
  interval = "months"
)
}
\arguments{
\item{x}{vector containing labels for x axis}

\item{foreground}{vector representing heights of bars visible in the foreground}

\item{background}{vector representing heights of bars visible in the background}

\item{triangles}{optional vector representing position of triangles}

\item{titles}{vector of series titles. Consists of 2 or 3 elements}

\item{styles}{optional dataframe of styles. First column contains styles for foreground series, second for background, third for triangles. dim(styles) must be length(x), length(titles)}
}
\value{
SVG string containing chart
}
\description{
Generate grouped column chart for visualizing up to 3 data series
}
