% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_chart_functions.R
\name{column_chart_reference}
\alias{column_chart_reference}
\title{Generate column chart with reference line.}
\usage{
column_chart_reference(
  df,
  x,
  series,
  ref_value,
  ref_label = NULL,
  styles = NULL
)
}
\arguments{
\item{df}{data frame in wide format containing data to be plotted}

\item{x}{vector containing labels for x axis or name of column in df with values of x axis labels}

\item{series}{vector containing name of column in df with values to plot}

\item{ref_value}{one element numeric vector with referencing value.}

\item{ref_label}{name of the referencing value}

\item{styles}{optional vector with styles of the bars}
}
\value{
SVG string containing chart
}
\description{
Generate column chart with reference line.
}
\examples{
#' # prepare some data frame
df <- data.frame(x = month.abb[1:6],
                 y = c(2, 4, 2, 1, 2.5, 3),
                 z = c(3, 4.5, 2, 1, 4, 2))

# generate character vector with svg data
svg <- column_chart_reference(df, x = 'x',
                              series = 'y',
                              ref_value = 3,
                              ref_label = 'baseline')

# show the plot
svg \%>\% SVGrenderer()
}
