% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_chart_functions.R
\name{column_chart_relative_variance}
\alias{column_chart_relative_variance}
\title{Generate column chart with relative variance (in percents)}
\usage{
column_chart_relative_variance(
  x,
  baseline,
  real,
  colors = 1,
  x_title,
  styles = NULL,
  interval = "months"
)
}
\arguments{
\item{x}{vector containing labels for x axis}

\item{baseline}{vector containing base values}

\item{real}{vector containing values that will be compared to baseline}

\item{colors}{1 if green color represents positive values having good buisness impact and red negative values having bad impact or 2 if otherwise}

\item{x_title}{the title of the plot}

\item{styles}{optional data frame of styles applied to pin heads}
}
\value{
SVG string containing chart
}
\description{
Generate column chart with relative variance (in percents)
}
\examples{
x <- month.abb
baseline <- rnorm(12, mean = 1, sd = 0.2)
real <- c(rnorm(6, mean = 0.8, sd = 0.2), rnorm(6, mean = 1.2, sd = 0.2))
column_chart_relative_variance(x, baseline, real, x_title = 'profit \%') \%>\%
  SVGrenderer()
}
