% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_chart_functions.R
\name{column_chart_waterfall}
\alias{column_chart_waterfall}
\title{Generate column waterfall chart for visualizing contribution.}
\usage{
column_chart_waterfall(df, x, series, styles = NULL)
}
\arguments{
\item{df}{data frame in wide format containing data to be plotted}

\item{x}{vector containing labels for x axis or name of column in df with values of x axis labels}

\item{series}{vector containing name of column in df with values to plot}

\item{styles}{optional vector with styles of the bars}
}
\value{
SVG string containing chart
}
\description{
Generate column waterfall chart for visualizing contribution.
}
