% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing_utils.R
\name{draw_triangle}
\alias{draw_triangle}
\title{Draw triangle and append it to svg string}
\usage{
draw_triangle(
  svg_string,
  tip_position_x,
  tip_position_y,
  orientation = "left",
  style = NULL,
  translate_vec = c(0, 0)
)
}
\arguments{
\item{svg_string}{svg string to paste a triangle}

\item{tip_position_x, tip_position_y}{x, y position of tip of the triangle}

\item{orientation}{where the triangle should be pointing. One of c('top', 'right', 'bottom', 'left').}

\item{style}{style of the triangle}

\item{translate_vec}{the translation vector}
}
\value{
svg string
}
\description{
Draw triangle and append it to svg string
}
