% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_scatter_chart.R
\name{factor_scatter_chart}
\alias{factor_scatter_chart}
\title{Use this function to create chart with a factor variable in y-axis,
a numeric variable in x-axis and an optional color variable}
\usage{
factor_scatter_chart(
  plot_data,
  x_name,
  y_name,
  color_name = NULL,
  static_color = "#4dbd5b",
  highlight = NULL,
  hover_name = NULL,
  plot_height = NULL,
  plot_width = NULL,
  show_legend = TRUE,
  show_x_axis_grid = FALSE,
  show_y_axis_grid = TRUE,
  marker_size = marker_default_size,
  y_text_wrap = 60,
  plot_title = NULL,
  x_axis_title = NULL,
  y_axis_title = NULL
)
}
\arguments{
\item{data.frame}{plot_data        This is the data.frame which will be used for the plot}

\item{char_vactor}{highlight        A vector of values from y_name which will be highlighted. Optional, if not provided every y-axis variable will have same color}

\item{Column}{hover_name       The column name of the variable that will be present in the hovers, can also be a normal vector.}

\item{boolean}{show_y_axis_grid This will let you display or hide the grids in the y-axis, default shown}

\item{num}{y_text_wrap      This will break the y-axis text by the number of characters specified, defaults to 60 characters}

\item{string}{y_axis_title     This is the y-axis title of the plot, defaults to NULL}
}
\value{
plot                         The plotly plot object as plot which can be viewed in HTML
}
\description{
Use this function to create chart with a factor variable in y-axis,
a numeric variable in x-axis and an optional color variable
}
\examples{

This will load two data.frames `marks_data` and `gender_school_earnings`
data("tidychartsdata")

You can plot by just using a data.frame, x-axis and y-axis.
factor_scatter_chart(marks_data, marks, name)

You can also change the color of the plot by specifying `static_color`
factor_scatter_chart(marks_data, marks, name, static_color = "#705bf5")

You can pass a third variable to get a plot with the color axis
factor_scatter_chart(marks_data, marks, name, subject)

}
