% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gender_school_earnings.R
\docType{data}
\name{gender_school_earnings}
\alias{gender_school_earnings}
\title{Sample data to test the graphs.}
\format{
gender_school_earnings is data frame with 21 observations and 4 variables:
\describe{
  \item{School}{The school where this data was collected}
  \item{Women}{Average earnings made by Women from that School}
  \item{Men}{Average earnings made by Men from that School}
  \item{Gap}{The gap between earnings of Men and Women for the respective School}
}
}
\source{
\url{https://raw.githubusercontent.com/plotly/datasets/master/school_earnings.csv}
}
\usage{
gender_school_earnings
}
\description{
Loading data('tidychartsdata') will attach 2 data.frames to your environment for testing out the graphs,
You can run the examples from the function documentation
}
\keyword{datasets}
