% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_functions.R
\name{get_color_stacked}
\alias{get_color_stacked}
\title{Function to get bar/area color for stacked plots.}
\usage{
get_color_stacked(series_number, colors_df = pkg.env$colors_df)
}
\arguments{
\item{series_number}{what is the number of the series. one of 1:6.}

\item{colors_df}{data frame with variety of colors}
}
\value{
list with bar_color and text_color
}
\description{
Function to get bar/area color for stacked plots.
}
