% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing_utils.R
\name{get_vector}
\alias{get_vector}
\title{Helper function to get the vector or column form df. If vector is passed it returns it.
If name of column is passed, it returns the column as a vector.}
\usage{
get_vector(df, vec)
}
\arguments{
\item{df}{data frame with a column}

\item{vec}{name of the column in df or vector of values}
}
\value{
vector
}
\description{
Helper function to get the vector or column form df. If vector is passed it returns it.
If name of column is passed, it returns the column as a vector.
}
