% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_plots.R
\name{join_plots}
\alias{join_plots}
\title{Join SVG plots.}
\usage{
join_plots(..., nrows = length(...), nclos = 1)
}
\arguments{
\item{...}{mumltiple character vectors with SVG content}

\item{nrows}{number of rows of plots in joint plot, default is set to number of plots}

\item{nclos}{number of columns of plots in joint plot, default is set to 1}
}
\value{
Character vector with SVG content
}
\description{
Join SVG plots.
}
\examples{
  df <- data.frame(
   mon = month.abb[1:6],
   values = rnorm(6)
  )

join_plots(
  column_chart(df, x = 'mon', series = 'values'),
  column_chart(df, x = 'mon', series = 'values')
) \%>\% SVGrenderer()
}
