% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidychart.R
\name{knit_print.tidychart}
\alias{knit_print.tidychart}
\title{Printing in knitr reports}
\usage{
\method{knit_print}{tidychart}(x, ...)
}
\arguments{
\item{x}{object of class tidychart to display in knitr document}

\item{...}{arguments passed to `knit_print` function}
}
\value{
object of class `html_screenshot` or `knit_asis`
}
\description{
Normally you don't want to use this function explicitly.
It is called automatically when printing output in knitr.
}
