% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot_with_few_points.R
\name{line_plot}
\alias{line_plot}
\title{Generates line plot with markers on every value.}
\usage{
line_plot(data, cat, series, series_labels, interval = "months", styles = NULL)
}
\arguments{
\item{data}{data frame containing data to be plotted}

\item{cat}{vector cointaining time interwals of the values}

\item{series}{vector containing names of columns in data with values to plot}

\item{series_labels}{vector containing names of series to be shown on the plot}

\item{interval}{intervals on x axis. The width of the bars depends on this parameter}
<<<<<<< HEAD
=======

\item{styles}{optional data frame with style names. Styles of the markers will be plotted accordingly.}
>>>>>>> e370a88fda030200c011417a8be48d315b3114dc
}
\value{
SVG string containing chart
}
\description{
Generates line plot with markers on every value.
}
