% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot_with_few_points.R
\name{line_plot_index}
\alias{line_plot_index}
\title{Generates line plot with markers on every value with index on a given value.}
\usage{
line_plot_index(
  data,
  cat,
  series,
  series_labels,
  index_val,
  interval = "months"
)
}
\arguments{
\item{data}{data frame containing data to be plotted}

\item{cat}{vector cointaining time interwals of the values}

\item{series}{vector containing names of columns in data with values to plot}

\item{series_labels}{vector containing names of series to be shown on the plot}

\item{index_val}{numeric value of the index}

\item{interval}{intervals on x axis. The width of the bars depends on this parameter}
}
\value{
SVG string containing chart
}
\description{
Generates line plot with markers on every value with index on a given value.
}
