% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot_with_many_points.R
\name{line_plot_many_points}
\alias{line_plot_many_points}
\title{Generates a line plot with markers on chosen points. Allows only one point per time interval. To create a plot with many points within one time interval try line_plot_with_many_points_complex().}
\usage{
line_plot_many_points(
  data,
  cat,
  series,
  series_labels,
  ser_names,
  point_cords,
  interval = "months"
)
}
\arguments{
\item{data}{data frame containing data to be plotted}

\item{cat}{vector cointaining time interwals of the values}

\item{series}{vector containing names of columns in data with values to plot}

\item{series_labels}{vector containing names of series to be shown on the plot}

\item{ser_names}{vector containing column names of a value to be marked}

\item{point_cords}{vector of the same length as ser_names containing numerical values of indexes in data of values to be marked}
}
\value{
SVG string containing chart
}
\description{
Generates a line plot with markers on chosen points. Allows only one point per time interval. To create a plot with many points within one time interval try line_plot_with_many_points_complex().
}
