% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot_many_points_complex.R
\name{line_plot_many_points_complex}
\alias{line_plot_many_points_complex}
\title{Generates a line plot with markers on chosen points, many points within one time interval allowed.}
\usage{
line_plot_many_points_complex(
  list,
  vector_x,
  vector_y,
  vector_cat,
  series_labels,
  df_numbers,
  point_cords
)
}
\arguments{
\item{list}{list of data frames, each representing one series. Data frame should consist of columns:
* containing numeric values from 0 to 100 defining the percentage of distance in one time interval of the point (x - coordinattes of the point)
* containing the value of a point  (y - coordinates of the point)
* containing the time interval of the value}

\item{vector_x}{vector containing the names of columns with x - coordinates of the point in the data frames}

\item{vector_y}{vector containing the names of columns with y - coordinates of the point in the data frames}

\item{vector_cat}{vector containing the names of columns with time interval of the point in the data frames}

\item{series_labels}{vector containing names of series to be shown on the plot}

\item{df_numbers}{vector containing index of data frame in the list of a value to be marked}

\item{point_cords}{vector of the same length as df_numbers containing numerical values of indexes in data frame of values to be marked}
}
\value{
SVG string containing chart
}
\description{
Generates a line plot with markers on chosen points, many points within one time interval allowed.
}
