% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot_normalized.R
\name{line_plot_normalized}
\alias{line_plot_normalized}
\title{Generates normalized areas (stacked lines) plot. If more than one series is supplied, stacked areas plot is generated.}
\usage{
line_plot_normalized(
  data,
  cat,
  series,
  series_labels,
  show_labels,
  interval = "months"
)
}
\arguments{
\item{data}{data frame containing data to be plotted}

\item{cat}{vector cointaining time interwals of the values}

\item{series}{vector containing names of columns in data with values to plot}

\item{series_labels}{vector containing names of series to be shown on the plot}

\item{show_labels}{vector of the same length as cat containg NA or not NA values defining which categories should have labels of values displayed}
}
\description{
Generates normalized areas (stacked lines) plot. If more than one series is supplied, stacked areas plot is generated.
}
