% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot_stacked.R
\name{line_plot_stacked}
\alias{line_plot_stacked}
\title{Generates areas (stacked lines) plot. If more than one series is supplied, stacked areas plot is generated.}
\usage{
line_plot_stacked(data, cat, series, series_labels, show_labels)
}
\arguments{
\item{data}{data frame containing data to be plotted}

\item{cat}{vector cointaining time interwals of the values}

\item{series}{vector containing names of columns in data with values to plot}

\item{series_labels}{vector containing names of series to be shown on the plot}

\item{show_labels}{vector of the same length as cat containg NA or not NA values defining which categories should have labels of values displayed}
}
\value{
SVG string containing chart
}
\description{
Generates areas (stacked lines) plot. If more than one series is supplied, stacked areas plot is generated.
}
