% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_marks_data.R
\docType{data}
\name{marks_data}
\alias{marks_data}
\title{Sample data to test the graphs.}
\format{
marks_data is data frame with 30 observations and 3 variables:
\describe{
  \item{name}{Name of the person}
  \item{subject}{Hypothetical subjects that they studied!}
  \item{marks}{The marks they might have secured if they weren't interrupted by the White Walkers}
}
}
\usage{
marks_data
}
\description{
Loading data('tidychartsdata') will attach 2 data.frames to your environment for testing out the graphs,
You can run the examples from the function documentation
}
\keyword{datasets}
