% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_scatter_chart.R
\name{numeric_scatter_chart}
\alias{numeric_scatter_chart}
\title{Use this function to create scatter plot with both x-axis and y-axis containing numeric variables
and with an optional color variable which is a factor variable}
\usage{
numeric_scatter_chart(
  plot_data,
  x_name,
  y_name,
  color_name = NULL,
  static_color = "#4dbd5b",
  hover_name = NULL,
  plot_height = NULL,
  plot_width = NULL,
  show_legend = TRUE,
  show_x_axis_grid = FALSE,
  show_y_axis_grid = FALSE,
  marker_size = marker_default_size,
  plot_title = NULL,
  x_axis_title = NULL,
  y_axis_title = NULL
)
}
\arguments{
\item{data.frame}{plot_data        This is the data.frame which will be used for the plot}

\item{Column}{hover_name       The column name of the variable that will be present in the hovers, can also be a normal vector.}

\item{boolean}{show_y_axis_grid This will let you display or hide the grids in the y-axis, default hidden}

\item{num}{marker_size      This is the size of the markers in the scatter plot, defaults to marker_default_size from constants.R}

\item{string}{y_axis_title     This is the y-axis title of the plot, defaults to NULL}
}
\value{
plot                         The plotly plot object as plot which can be viewed in HTML
}
\description{
Use this function to create scatter plot with both x-axis and y-axis containing numeric variables
and with an optional color variable which is a factor variable
}
\examples{

You can plot by just using a data.frame, x-axis and y-axis.
numeric_scatter_chart(iris, Sepal.Length, Petal.Length)

You can also change the color of the plot by specifying `static_color`
numeric_scatter_chart(iris, Sepal.Length, Petal.Length, static_color = "#f55b96")

You can pass a third variable to get a plot with the color axis
numeric_scatter_chart(iris, Sepal.Length, Petal.Length, color_name = Species)

}
