% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_functions.R
\name{restore_defaults}
\alias{restore_defaults}
\title{Restore default color and style settings}
\usage{
restore_defaults()
}
\value{
No return value, called for side effects
}
\description{
Restore default color and style settings
}
\examples{

restore_defaults()
}
