% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_and_bubble_plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Generates a scatter plot. If additional argument added, a bubble plot is generated.}
\usage{
scatter_plot(
  data,
  x,
  y,
  cat,
  x_space_size,
  y_space_size,
  x_names,
  y_names,
  legend_title,
  bubble_value = NULL
)
}
\arguments{
\item{data}{data frame containing data to be plotted}

\item{x}{vector containing x - coordinates of values}

\item{y}{vector containing y - coordinates of values}

\item{cat}{vector cointaining categories of the values}

\item{x_space_size}{numeric value of the space between the ticks on the x- axis}

\item{y_space_size}{numeric value of the space between the ticks on the y- axis}

\item{x_names}{vector containing two values:
* name of the value presented on the x - axis
* units of values presented on the x - axis}

\item{y_names}{vector containing two values:
* name of the value presented on the y - axis
* units of values presented on the y - axis}

\item{legend_title}{title of the legend}

\item{bubble_value}{vector containing values defining the size of bubbles. Set by default to NULL.}
}
\description{
Generates a scatter plot. If additional argument added, a bubble plot is generated.
}
