% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_functions.R
\name{set_colors}
\alias{set_colors}
\title{Change default colors of the package}
\usage{
set_colors(colors_df)
}
\arguments{
\item{colors_df}{data frame with 6 rows and 2 columns. Columns must nave names : "text_colors", "bar_colors"}
}
\description{
Change default colors of the package
}
