% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_functions.R
\name{set_margins}
\alias{set_margins}
\title{Set margins size.}
\usage{
set_margins(x = NULL, ...)
}
\arguments{
\item{x}{A named list with numeric margin values. Default set to NULL.}

\item{...}{Key - value pairs, where key is the name of margin and value is a numeric value of a margin.}
}
\value{
No return value, called for side effects.
}
\description{
Currently supported is setting 'top' and 'left' margins. Names of elements in x and keys in ... should be 'top' or 'left'.
}
\examples{
set_margins(list(top = 75, left = 80))

set_margins(top = 75, left = 80)
}
