% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_functions.R
\name{set_scatter_colors}
\alias{set_scatter_colors}
\title{Change default colors of the scatter plots from the package.}
\usage{
set_scatter_colors(new_scatter_colors)
}
\arguments{
\item{new_scatter_colors}{vector of the length of 6 containing rgb values of chosen colors in format: "rgb(x,y,z)"}
}
\value{
No return value, called for side effects.
}
\description{
Customize your scatter plots and change default color palette.
}
\examples{
mi2lab_scatter_colors <- c(
   "rgb(68, 19, 71)",
   "rgb(243, 46, 255)",
   "rgb(106, 0, 112)",
   "rgb(217, 43, 227)" ,
   "rgb(114, 49, 117)",
   "rgb(249, 110, 255)"
 )

set_scatter_colors(mi2lab_scatter_colors)

}
