% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_functions.R
\name{set_styles}
\alias{set_styles}
\title{Change default styles for plots}
\usage{
set_styles(styles_df)
}
\arguments{
\item{styles_df}{data frame with columns 'fill' and 'stroke'. Rows represent subsequent styles which names can be passed to plotting functions, usually as styles argument.}
}
\description{
Change default styles for plots
}
