% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_utils.R
\name{show}
\alias{show}
\title{Function to render SVG as bitmap, used only in tests}
\usage{
show(svg_string)
}
\arguments{
\item{svg_string}{string containing SVG statements}
}
\value{
No return value, called for side effects
}
\description{
Function to render SVG as bitmap, used only in tests
}
\examples{
if(FALSE){

  df <- data.frame(x = c(1,2,3), y = c(4,5,6))

  column_chart(df, x = df$x, series = c("y")) \%>\% show()
}

}
