% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing_utils.R
\name{str_width}
\alias{str_width}
\title{Caclulate string width in pixels}
\usage{
str_width(string, bold = FALSE)
}
\arguments{
\item{string}{string which width will be calculated}

\item{bold}{boolean value, if string will be written in bold}
}
\value{
string width in pixels
}
\description{
Caclulate string width in pixels
}
