% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_utils.R
\name{SVGsave}
\alias{SVGsave}
\title{Save svg image.}
\usage{
SVGsave(svg_string, path)
}
\arguments{
\item{svg_string}{string containing SVG statements}

\item{path}{path to file where image will be saved}
}
\value{
No return value, called for side effects
}
\description{
Save svg image.
}
