% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart_plot.R
\name{barchart_plot_normalized}
\alias{barchart_plot_normalized}
\title{Generates normalized horizontal barchart. If more than one series is supplied, stacked barchart is generated.}
\usage{
barchart_plot_normalized(data, cat, series, series_labels)
}
\arguments{
\item{data}{data frame containing data to be plotted}

\item{cat}{vector cointaining category names of values}

\item{series}{vector containing names of columns in data with values to plot}

\item{series_labels}{vector containing names of series to be shown on the plot}
}
\value{
SVG string containing chart
}
\description{
Generates normalized horizontal barchart. If more than one series is supplied, stacked barchart is generated.
}
