% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_chart_functions.R
\name{column_chart_waterfall_variance}
\alias{column_chart_waterfall_variance}
\title{Generate column waterfall chart with absolute variance}
\usage{
column_chart_waterfall_variance(
  x,
  baseline,
  real,
  colors = 1,
  data = NULL,
  result_title,
  interval = "months"
)
}
\arguments{
\item{x}{vector containing labels for x axis or name of column in data with values of x axis labels}

\item{baseline}{vector containing base values or name of column in data with base values}

\item{real}{vector containing values that will be compared to baseline  or name of column in data with that values}

\item{colors}{1 if green color represents positive values having good buisness impact and red negative values having bad impact or 2 if otherwise}

\item{data}{data frame with columns containing data for x, baseline or real series}

\item{result_title}{title for the result bar}

\item{interval}{intervals on x axis. The width of the bars depends on this parameter}
}
\value{
SVG string containing chart
}
\description{
Generate column waterfall chart with absolute variance
}
\examples{
x <- month.abb
baseline <- rnorm(12)
real <- c(rnorm(6, mean = -1), rnorm(6, mean = 1))
column_chart_waterfall_variance(x, baseline, real, result_title = 'year profit') \%>\%
  SVGrenderer()
}
