% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumbbell_chart.R
\name{dumbbell_chart}
\alias{dumbbell_chart}
\title{Use this function to create a dumbbell chart to compare two numeric variables along a factor variable}
\usage{
dumbbell_chart(
  plot_data,
  x1_name,
  x2_name,
  y_name,
  line_color = "#bdbdbd",
  x1_color = "#5884d6",
  x2_color = "#bbe64e",
  plot_height = NULL,
  plot_width = NULL,
  show_legend = TRUE,
  show_x_axis_grid = FALSE,
  show_y_axis_grid = FALSE,
  marker_size = marker_default_size,
  plot_title = NULL,
  x_axis_title = NULL,
  y_axis_title = NULL
)
}
\arguments{
\item{data.frame}{plot_data        This is the data.frame which will be used for the plot}

\item{Column}{y_name           The column name of the the variable in the y-axis, It has to be a factor variable}

\item{boolean}{show_y_axis_grid This will let you display or hide the grids in the y-axis, default hidden}

\item{num}{marker_size      This is the size of the markers in the scatter plot, defaults to marker_default_size from constants.R}

\item{string}{y_axis_title     This is the y-axis title of the plot, defaults to NULL}
}
\value{
plot                         The plotly plot object as plot which can be viewed in HTML
}
\description{
Use this function to create a dumbbell chart to compare two numeric variables along a factor variable
}
\examples{

This will load two data.frames `marks_data` and `gender_school_earnings`
data("tidychartsdata")

You can plot a dumbbell chart using by just specifying data, x1_name, x2_name and y_name.
dumbbell_chart(gender_school_earnings, Men, Women, School)

If you want to specify particular color, you can specify them to their respective arguments
dumbbell_chart(gender_school_earnings, Men, Women, School, line_color = "#6ba1e8", x1_color = "#4c7ced", x2_color = "#a7d1eb")

}
