% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot_with_few_points.R
\name{line_plot}
\alias{line_plot}
\title{Generates line plot with markers on every value.}
\usage{
line_plot(data, cat, series, series_labels, interval = "months")
}
\arguments{
\item{data}{data frame containing data to be plotted}

\item{cat}{vector cointaining time interwals of the values}

\item{series}{vector containing names of columns in data with values to plot}

\item{series_labels}{vector containing names of series to be shown on the plot}
}
\value{
SVG string containing chart
}
\description{
Generates line plot with markers on every value.
}
