% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot_many_points_complex.R
\name{line_plot_many_points_wrapper}
\alias{line_plot_many_points_wrapper}
\title{Title}
\usage{
line_plot_many_points_wrapper(df, dates, series, scale = "months")
}
\arguments{
\item{scale}{}
}
\value{

}
\description{
Title
}
