% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_functions.R
\name{set_styles}
\alias{set_styles}
\title{Change default styles for plots}
\usage{
set_styles(styles_df)
}
\arguments{
\item{styles_df}{data frame with columns 'fill' and 'stroke'. Rows represent subsequent styles which names can be passed to plotting functions, usually as styles argument.}
}
\value{
No return value, called for side effects
}
\description{
Change default styles for plots
}
\examples{
styles_df <-
  rbind(
  actual = c("rgb(64,64,64)", "rgb(64,64,64)"),
  previous =
    c("rgb(166,166,166)", "rgb(166,166,166)"),
  forecast =
    c("url(#diagonalHatch)", "rgb(64,64,64)"),
  plan = c("white", "rgb(64,64,64)"),
  total_white = c("white", "white")
  )
colnames(styles_df) <- c("fill", "stroke")

set_styles(styles_df)
}
