% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGrenderer.R
\name{SVGrenderer}
\alias{SVGrenderer}
\title{Function to render SVG image as htmlwidget}
\usage{
SVGrenderer(svg_string, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{svg_string}{one element character vector with image in svg format}

\item{width}{width of the widget}

\item{height}{height of the widget}

\item{elementId}{HTML element ID}
}
\value{
No return value, called for side effects
}
\description{
Use this function to show SVG images from given string in SVG embedded in HTML.
}
